+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.2.13] - 2020-08-03
+
 ### Added
 - `actingAsJwtUser` helper for tests.